[Refactor] StringLengthUtil 분리, 닉네임 검사 로직 서비스->도메인 으로 변경
